Fix license for use on Android systems
Sorry for the bold change, but at the moment the actually prevents the library from its intended use. If the license restricts its use to be ` for use on ESPRESSIF SYSTEMS ESP8266/ESP32 only` it cannot be used to be run **on** an Android device.
Could you please fix this? As I lack any knowledge in Mandarin I am not able to make an suggestion for the rest of the file.